refactor: click handler into hook
