+++ code/system.ui/react.common/src/ui/PropList/-dev/-SPEC.Sample.Builder.tsx
